+++ bluemog-client/src/components/molecules/Image/index.stories.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import ProfileImage, { Props } from '.';
+import Image, { Props } from '.';
 import imageFile from '../../../../public/assets/images/gummy.jpg';
 
 export default {
-  component: ProfileImage,
-  title: 'Components/Molcules/Profile Image',
+  component: Image,
+  title: 'Components/Molecules/Image',
 };
 
-const Template = (args: Props) => <ProfileImage {...args} />;
+const Template = (args: Props) => <Image {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {
